Rename Trie to HexaryTrie
