+++ tests/test_proof.py
@@ -1,25 +1,25 @@
 import pytest
 
-from trie import Trie
+from trie.hexary import HexaryTrie
 from trie.exceptions import BadTrieProof
 from trie.utils.sha3 import keccak
 
 
 def test_get_from_proof_key_exists():
     from .sample_proof_key_exists import key, state_root, proof
-    assert Trie.get_from_proof(state_root, key, proof) != b''
+    assert HexaryTrie.get_from_proof(state_root, key, proof) != b''
 
 
 def test_get_from_proof_key_does_not_exist():
     from .sample_proof_key_does_not_exist import key, state_root, proof
-    assert Trie.get_from_proof(state_root, key, proof) == b''
+    assert HexaryTrie.get_from_proof(state_root, key, proof) == b''
 
 
 def test_get_from_proof_invalid():
     from .sample_proof_key_exists import key, state_root, proof
     proof[5][3] = b''
     with pytest.raises(BadTrieProof):
-        Trie.get_from_proof(state_root, key, proof)
+        HexaryTrie.get_from_proof(state_root, key, proof)
 
 
 def test_get_from_proof_empty():
@@ -27,4 +27,4 @@ def test_get_from_proof_empty():
     key = keccak(b'some key')
     proof = []
     with pytest.raises(BadTrieProof):
-        Trie.get_from_proof(state_root, key, proof)
+        HexaryTrie.get_from_proof(state_root, key, proof)
